requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
data==0.4
decorator==5.1.1
fonttools==4.38.0
fred==3.1
fredapi==0.5.0
frozendict==2.3.5
funcsigs==1.0.2
html5lib==1.1
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
load-dotenv==0.1.0
lxml==4.9.2
matplotlib==3.6.3
multitasking==0.0.11
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
Pillow==9.4.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
webencodings==0.5.1
yfinance==0.2.10
zipp==3.14.0
